requirements.txt
requests==2.22.0
tweepy==3.10.0
flask==1.1.1
cairosvg==2.5.1
pytz==2020.5
numpy==1.19.5
sentry-sdk==0.19.5
google-cloud-storage==1.35.0
glob2==0.7
Pillow==8.1.0